Fix PostMaster (if previously installed)
Python is not decompressed by default on CrossMix (a lot of small files), the PortMaster install should do it on first boot BUT if you have already installed PortMaster before (with another OS from CrossMix) it will not decompress Python again.
Removing the python3 will force systematically the PortMaster install on first boot even if it has already been installed previously in the internal storage.