Update PPSSPP with hotfix 9
PSP folder reorganized, 3 versions of PPSSPP:

- PPSSPP 1.17.1 - OpenGL
- PPSSPP 1.17.1 - Vulkan
- PPSSPP-1.5.4

Hotfix 9 observations:
The 1.5.4 has not changed since hotfix 6.
There are a little difference in the vulkan binary compared to hotfix 6.
It now includes a GL version of the 1.17.1 (however it's potentially possible to use GL with the vulkan binary if you set the right option in ppsspp.ini).

1.17.1 - OpenGL &  Vulkan now share the same config folder, it allows to avoid to remake the game configuration each time you switch between GL and Vulkan. Also the save states are mutualized between both version.

The performance mode is enabled by detection of the label which have been selected in launcher menu, a script which could be useful for other use cases. It avoid to have a lot of duplicated launcher scripts.