Add emu_intro.sh
This script will be used to triggers events when an emulator is selected.
Currently it runs "effect.sh" from the current Emu folder and if an sound file exists for this emulator it plays it.
Path of the emulator intro sound file:  "theme/{Current Theme}/sound/intro_{Current Emulator Path}.wav"