Add a launch script for PICO-8 Splore
Splore game store from PICO-8 can be launched without games so it was necessary to add a fake rom to be able to run Splore from MainUI. As for ScummVM, the fake rom file extension is used as flag to know if the user has already launched Splore or not. Once the flag modified EmuCleaner will not suppress anymore the emulator even if there are no other roms.

How to use Splore from PICO-8 Wrapper on CrossMix-OS :
-----------------------------------------------------------------
- To use Pico-8 Wrapper and Splore you need official purchased PICO-8 binaries (details here : https://github.com/cizia64/CrossMix-OS/tree/main/Emus/PICO/PICO8_Wrapper)
- Go in emulators list, press Menu button and select "Edit"
- Select PICO-8 emulator in the checklist
- Then go in PICO-8 rom list and launch "° Run Splore"
- Done !