Add Region setting for emulator labels
- add "emulators_list.db" database which contains  directory names, emulator names (depending the region), manufacturer, release date, ... It will be a database to maintain to allow quick renaming via System Tools.
- add scripts which allow to rename all the emulators in one click

This database will be used to offer the possibility to the user to select the kind of emulator label he want:
https://docs.google.com/spreadsheets/d/1sSi6PamSL6nHU3nORooms9AjNKYcoGGDZNJbUkF-sa0/edit?usp=sharing
For example US users want "Genesis" and EU want "Mega Drive". Do not hesitate to post an issue if there are errors.

For each US/EU/JP there is a full name and a short name. The short name will be used to prefix with manufacturer name (and group the emulators by manufacturer).

For example you'll have the choice between:
- if you're from EU: Mega Drive or Seg. MD
- if you're from US: Genesis or Seg. Gen.

The current type of emulator label is kept in crossmix.json as "EMU LABELS".

These emulator rename scripts will be used later by CrossMix "Theme Pack Selector".