Workaround for ports controls problems
This is really a strange behavior from the stock firmware: Input devices are not "hotplug": I have remarked that if you plug any input device when an app is already running, the input device will not be detected.
For example on MainUI (or any other app like a port), plug an USB keyboard: it doesn't work until you restart MainUI. The problem is not here on Knulli.  gptokeyb creates virtual input devices often after the launch of the port so it's the same problem.