[Emus] Add GZDoom standalone (#263)
GZDoom v4.13.2
----------------------
GZDoom is a modern port of Doom. It also supports Heretic, Hexen, Strife, Chex Quest, and fan-created games like Harmony and Hacx. On top of that mods are supported too.
The CrossMix default configuration is inspired from recommended settings from advanced users of the GZDoom community.

A quick sum up
---------------------
- iwad files are main game files. Doom 1, Doom 2, Hexen... each one has its one iwad.
iwad extensions are generally .wad (😅) and must be placed in /mnt/SDCARD/Roms/DOOM/GZDoom
- mod /patch files can be wad ,pk3, ipk3, pk7... They modify the behavior, textures, weapons of the previsouly selected iwad

Usage
--------
The CrossMix implementation is made to facilitate the use of GZDoom.

Copy your iwads in /mnt/SDCARD/Roms/DOOM/GZDoom
Copy your mods/patches in /mnt/SDCARD/Roms/DOOM/GZDoom/mods

Go in Emulators -> DOOM -> GZDoom and run the iwad you have added (if you don't see it remember to run EmuCleaner app and refresh roms)
If you have mods available, you'll be prompted by a file selector screen (press B if you don't want to load any mod for your previously selected iwad).

CrossMix includes an iwad checksum checker, it will tell you if the iwad is not recognized (which is not a good sign for using it with GZDoom) and will give you the current version of your iwad if it is recognized.
Credits: https://github.com/Ch0wW/wadverifier  for the iwad checksums database et son article de blog:
https://blog.ch0ww.fr/iwadverifier/

How to find iwads
------------------------
The list of supported iwad/games for GZDoom is located here: https://zdoom.org/wiki/IWAD
Most of the games are licenced games, so it requires to add your own files (these old Doom games are generally available on all game online platforms for few dollars).
Note that free shareware versions exist and give a very good first experience of the capacities of GZDoom.
A pack of the Free-to-play games listed on zdoom wiki is available here (just extract it on sd card root):
https://drive.google.com/file/d/1zk4h7ommRgY08dI31rBkpdN4O3EC5pps/view?usp=drive_link

How to find mods
------------------------
There is a list on the official zdoom wiki:
https://zdoom.org/wiki/Category:WADs_created_for_ZDoom

DoomLauncher is a launcher for Windows which allows to download mods, and test it with GZDoom for Windows directly. It's a good way to test the compatibility of your mod like this before moving it your handheld device.
https://github.com/nstlaurent/DoomLauncher