Adding lcd control scripts and replacing toggle_lcd.sh (#391)
New scripts to turn the lcd on, off and toggle it. 
It should be a bit easier to interact with the display and do things like turn it off during ssh sessions :

export PATH="/mnt/SDCARD/System/bin:/mnt/SDCARD/System/usr/trimui/scripts:$PATH"
lcd_off.sh
lcd_on.sh
lcd_toggle.sh

When turning the display ON, it will set the brightness to whatever was configured in the Settings, instead of setting it to full brightness.