chore: new fn_editor option: turn on led
Adding the option to have an fn_editor command that turns the led on when the fn_button is on.
This is useful to highligh when the fn_button is actually running.
Do notice that led_on.sh just like its counterpart, do not respect whatever the led "controls" have defined.
I understand that is a bug that can be solved later on.